update polymorphic owner links for ownable resources
