+++ app/serializers/acl_serializer.rb
@@ -16,10 +16,10 @@ module ACLSerializer
       links = super
       links.delete("#{key}.user_group")
       link_type = resource_type.pluralize
-      links["#{key}.#{resource_type}"] = { type: link_type,
-                                           href: "/#{link_type}/{#{key}.#{resource_type}}" }
-      links["#{key}.owner"] = { type: "owners",
-                                href: "/{#{key}.owner.href}/{#{key}.owner.id}" }
+      links["#{key}.#{resource_type}"] = { href: "/#{link_type}/{#{key}.#{resource_type}}",
+                                           type: link_type }
+      links["#{key}.owner"] = { href: "/{#{key}.owner.href}/{#{key}.owner.id}",
+                                type: "owners" }
       links
     end
   end
