+++ app/serializers/collection_serializer.rb
@@ -1,5 +1,7 @@
 class CollectionSerializer
   include RestPack::Serializer
+  include OwnerLinkSerializer
+
   attributes :id, :name, :display_name, :created_at, :updated_at
   can_include :project, :owner
   can_filter_by :display_name, :name
