+++ spec/support/polymorphic_owner_links_serializer.rb
@@ -1,4 +1,4 @@
-shared_examples "an ACL roles serializer" do
+shared_examples "it has custom owner links" do
   it "should have the custom owner resource links" do
     get :index
     resource_links = json_response[api_resource_name].map do |resource|
@@ -11,7 +11,8 @@ shared_examples "an ACL roles serializer" do
     get :index
     owner_prefix = "#{api_resource_name}.owner"
     owner_link = json_response["links"][owner_prefix]
-    expected = {"type" => "owners", "href" => "/{#{owner_prefix}.href}/{#{owner_prefix}.id}"}
+    expected = { "href" => "/{#{owner_prefix}.href}/{#{owner_prefix}.id}",
+                 "type" => "owners" }
     expect(owner_link).to eq(expected)
   end
 end
